pfblockerng: apply review nitpicks (#1257, #1258)
- Extract assertNoPhpWarning() into PfbNoPhpWarningTrait, shared by
  LogTrimNeededTest and LogAgeTrimNeededTest. It was defined twice -- the same
  'one contract, two definitions' defect this PR exists to remove (Copilot).
- Read pfb_log_trim_margin_pct's raw config value once into $margin_raw instead
  of repeating the expression at both pfb_log_mgmt() call sites. Hoisting the
  READ is safe; hoisting the PARSE is what let an unclamped margin through.
- Shrink the bounded-memory fixture from 200 MiB to 32 MiB: still 32x any
  per-chunk buffer, so a slurp cannot hide in the headroom, without 200 MB of
  CI disk churn per suite run. (The stated runtime concern was measurably
  false -- the test ran in 0.21s -- but the disk churn was real.)
- Record in the short-read test that the !feof branch is defence-in-depth: a
  real local file's empty fread() always coincides with feof()===TRUE, so
  production reaches the fread()===FALSE branch instead.
