+++ tests/php/LogTrimNeededTest.php
@@ -2,6 +2,8 @@
 
 declare(strict_types=1);
 
+require_once __DIR__ . '/PfbNoPhpWarningTrait.php';
+
 use PHPUnit\Framework\Attributes\CoversFunction;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
@@ -24,6 +26,8 @@ use PHPUnit\Framework\TestCase;
 #[CoversFunction('pfb_log_trim_needed')]
 final class LogTrimNeededTest extends TestCase
 {
+	use PfbNoPhpWarningTrait;
+
 	private string $tmpFile;
 
 	protected function setUp(): void
@@ -41,23 +45,6 @@ final class LogTrimNeededTest extends TestCase
 	private function daysAgo(int $days): string
 	{
 		return date('Y-m-d H:i:s', time() - ($days * 86400));
-	}
-
-	/** Runs $fn under a handler that records E_WARNING instead of printing it; fails the test if any fired. */
-	private function assertNoPhpWarning(callable $fn): mixed
-	{
-		$warnings = [];
-		set_error_handler(static function (int $errno, string $errstr) use (&$warnings): bool {
-			$warnings[] = $errstr;
-			return TRUE;
-		}, E_WARNING);
-		try {
-			$result = $fn();
-		} finally {
-			restore_error_handler();
-		}
-		$this->assertSame([], $warnings, 'must not trigger a PHP warning: ' . implode('; ', $warnings));
-		return $result;
 	}
 
 	private function seedLines(int $count): void
